+++ src/SecurityCheckPlugin.php
@@ -91,15 +91,19 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	public const CODE_TAINT = 1 << 13;
 	public const CODE_EXEC_TAINT = 1 << 14;
 
+	// User-controlled regular expressions, for ReDoS
+	public const REGEX_TAINT = 1 << 15;
+	public const REGEX_EXEC_TAINT = 1 << 16;
+
 	// For stuff that doesn't fit another category
-	public const MISC_TAINT = 1 << 15;
-	public const MISC_EXEC_TAINT = 1 << 16;
+	public const MISC_TAINT = 1 << 17;
+	public const MISC_EXEC_TAINT = 1 << 18;
 
 	// To allow people to add other application specific taints.
-	public const CUSTOM1_TAINT = 1 << 17;
-	public const CUSTOM1_EXEC_TAINT = 1 << 18;
-	public const CUSTOM2_TAINT = 1 << 19;
-	public const CUSTOM2_EXEC_TAINT = 1 << 20;
+	public const CUSTOM1_TAINT = 1 << 19;
+	public const CUSTOM1_EXEC_TAINT = 1 << 20;
+	public const CUSTOM2_TAINT = 1 << 21;
+	public const CUSTOM2_EXEC_TAINT = 1 << 22;
 
 	// Special purpose for supporting MediaWiki's IDatabase::select
 	// and friends. Like SQL_TAINT, but only applies to the numeric
@@ -109,12 +113,12 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	// The associative keys also have this flag if they are tainted.
 	// It is also assumed anything with this flag will also have
 	// the SQL_TAINT flag set.
-	public const SQL_NUMKEY_TAINT = 1 << 21;
-	public const SQL_NUMKEY_EXEC_TAINT = 1 << 22;
+	public const SQL_NUMKEY_TAINT = 1 << 23;
+	public const SQL_NUMKEY_EXEC_TAINT = 1 << 24;
 
 	// For double escaped variables
-	public const ESCAPED_TAINT = 1 << 23;
-	public const ESCAPED_EXEC_TAINT = 1 << 24;
+	public const ESCAPED_TAINT = 1 << 25;
+	public const ESCAPED_EXEC_TAINT = 1 << 26;
 
 	// Special purpose flags (Starting at 2^28)
 	// Cancel's out all EXEC flags on a function arg if arg is array.
@@ -134,7 +138,8 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 
 	// YES_TAINT denotes all taint a user controlled variable would have
 	public const YES_TAINT = self::HTML_TAINT | self::SQL_TAINT | self::SHELL_TAINT | self::SERIALIZE_TAINT |
-		self::PATH_TAINT | self::CODE_TAINT | self::CUSTOM1_TAINT | self::CUSTOM2_TAINT | self::MISC_TAINT;
+		self::PATH_TAINT | self::CODE_TAINT | self::REGEX_TAINT | self::CUSTOM1_TAINT | self::CUSTOM2_TAINT |
+		self::MISC_TAINT;
 	public const EXEC_TAINT = self::YES_TAINT << 1;
 	public const YES_EXEC_TAINT = self::YES_TAINT | self::EXEC_TAINT;
 
@@ -313,6 +318,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 			self::CUSTOM2_TAINT => 'CUSTOM2',
 			self::CODE_TAINT => 'CODE',
 			self::PATH_TAINT => 'PATH',
+			self::REGEX_TAINT => 'REGEX',
 			self::MISC_TAINT => 'MISC',
 			self::SQL_NUMKEY_TAINT => 'SQL_NUMKEY',
 			self::ARRAY_OK => 'ARRAY_OK',
@@ -325,6 +331,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 			self::CUSTOM2_EXEC_TAINT => '*CUSTOM2',
 			self::CODE_EXEC_TAINT => '*CODE',
 			self::PATH_EXEC_TAINT => '*PATH',
+			self::REGEX_EXEC_TAINT => '*REGEX',
 			self::MISC_EXEC_TAINT => '*MISC',
 			self::SQL_NUMKEY_EXEC_TAINT => '*SQL_NUMKEY',
 		];
@@ -483,7 +490,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	 */
 	public static function parseTaintLine( string $line ) : ?Taintedness {
 		$types = '(?P<type>htmlnoent|html|sql|shell|serialize|custom1|'
-			. 'custom2|misc|code|path|sql_numkey|escaped|none|tainted)';
+			. 'custom2|misc|code|path|regex|sql_numkey|escaped|none|tainted)';
 		$prefixes = '(?P<prefix>escapes|onlysafefor|exec)';
 		$taintExpr = "/^(?P<taint>(?:${prefixes}_)?$types|array_ok|allow_override|raw_param)$/";
 
@@ -542,7 +549,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	 *
 	 * @note htmlnoent treated like self::HTML_TAINT.
 	 * @param string $name one of:
-	 *   html, sql, shell, serialize, custom1, custom2, code, path, misc, sql_numkey,
+	 *   html, sql, shell, serialize, custom1, custom2, code, path, regex, misc, sql_numkey,
 	 *   escaped, none (= self::NO_TAINT), tainted (= self::YES_TAINT)
 	 * @return int One of the TAINT constants
 	 */
@@ -565,6 +572,8 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 				return self::CODE_TAINT;
 			case 'path':
 				return self::PATH_TAINT;
+			case 'regex':
+				return self::REGEX_TAINT;
 			case 'misc':
 				return self::MISC_TAINT;
 			case 'sql_numkey':
@@ -587,6 +596,22 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	 * @phan-return array<string,int[]>
 	 */
 	protected function getPHPFuncTaints() : array {
+		$pregMatchTaint = [
+			self::REGEX_EXEC_TAINT,
+			self::YES_TAINT,
+			self::NO_TAINT, // TODO Possibly unsafe pass-by-ref,
+			self::NO_TAINT,
+			self::NO_TAINT,
+			'overall' => self::NO_TAINT,
+		];
+		$pregReplaceTaint = [
+			self::REGEX_EXEC_TAINT,
+			self::YES_TAINT, // TODO This is used for strings (in preg_replace) and callbacks (in preg_replace_callback)
+			self::YES_TAINT,
+			self::NO_TAINT,
+			self::NO_TAINT,
+			'overall' => self::NO_TAINT
+		];
 		return [
 			'\htmlentities' => [
 				self::ESCAPES_HTML,
@@ -752,6 +777,44 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 				self::HTML_EXEC_TAINT,
 				self::HTML_EXEC_TAINT,
 				self::HTML_EXEC_TAINT,
+				'overall' => self::NO_TAINT
+			],
+			'\preg_filter' => [
+				self::REGEX_EXEC_TAINT,
+				self::YES_TAINT,
+				self::YES_TAINT,
+				self::NO_TAINT,
+				self::NO_TAINT,
+				'overall' => self::NO_TAINT
+			],
+			'\preg_grep' => [
+				self::REGEX_EXEC_TAINT,
+				self::YES_TAINT,
+				self::NO_TAINT,
+				'overall' => self::NO_TAINT
+			],
+			'\preg_match_all' => $pregMatchTaint,
+			'\preg_match' => $pregMatchTaint,
+			'\preg_quote' => [
+				self::YES_TAINT & ~self::REGEX_TAINT,
+				self::NO_TAINT,
+				'overall' => self::NO_TAINT
+			],
+			'\preg_replace' => $pregReplaceTaint,
+			'\preg_replace_callback' => $pregReplaceTaint,
+			'\preg_replace_callback_array' => [
+				self::REGEX_EXEC_TAINT,
+				self::YES_TAINT,
+				self::NO_TAINT,
+				self::NO_TAINT,
+				self::NO_TAINT,
+				'overall' => self::NO_TAINT
+			],
+			'\preg_split' => [
+				self::REGEX_EXEC_TAINT,
+				self::YES_TAINT,
+				self::NO_TAINT,
+				self::NO_TAINT,
 				'overall' => self::NO_TAINT
 			]
 		];